adding back netstandard1.3 config
