+++ src/System.IO.Packaging/ref/Configurations.props
@@ -1,6 +1,7 @@
 ﻿<Project>
   <PropertyGroup>
     <PackageConfigurations>
+      netstandard1.3;
       net46;
       netstandard;
     </PackageConfigurations>
