+++ src/System.IO.Packaging/ref/System.IO.Packaging.PackUriScheme.cs
@@ -1,0 +1,18 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+// ------------------------------------------------------------------------------
+// Changes to this file must follow the http://aka.ms/api-review process.
+// ------------------------------------------------------------------------------
+
+namespace System.IO.Packaging
+{
+    public static partial class PackUriHelper
+    {
+        public static Uri GetPackageUri(System.Uri packUri) { throw null; }
+        public static System.Uri Create(System.Uri packageUri) { throw null;}
+        public static System.Uri Create(System.Uri packageUri, System.Uri partUri) { throw null; }
+        public static System.Uri Create(System.Uri packageUri, System.Uri partUri, string fragment) { throw null; }
+        public static int ComparePackUri(Uri firstPackUri, Uri secondPackUri) { throw null; }
+    }
+}
