+++ src/System.IO.Packaging/ref/System.IO.Packaging.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <ProjectGuid>{4DD91348-1D1F-4964-B5CB-047D916E79EE}</ProjectGuid>
     <IsPartialFacadeAssembly Condition="'$(TargetGroup)' == 'netfx' OR '$(TargetGroup)' == 'net46'">true</IsPartialFacadeAssembly>
@@ -10,6 +10,7 @@
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' != 'netstandard1.3'">
     <Compile Include="System.IO.Packaging.Serialization.cs" />
+    <Compile Include="System.IO.Packaging.PackUriScheme.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' == 'netfx' OR '$(TargetGroup)' == 'net46'">
     <Reference Include="mscorlib" />
