+++ src/System.IO.Packaging/src/Configurations.props
@@ -2,6 +2,7 @@
   <PropertyGroup>
     <PackageConfigurations>
       net46;
+      netstandard1.3;
       netstandard;
     </PackageConfigurations>
     <BuildConfigurations>
